requirements.txt
alembic==1.4.2
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Misaka==1.0.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
misaka==2.1.1
mysql-client==0.0.1
mysqlclient==1.4.6
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.17
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
